Code for Day 4 puzzle Here we go again!  Another data class. I write very verbose code, I'm not sure how I feel about that. Spent 20min writting Part1, then spent about an hour trying to find my mistake.
I wrote:
if (!isset($aggregate['@']) && $aggregate['@'] < 4)

Rather than:
if (!isset($aggregate['@']) || $aggregate['@'] < 4)

I was so used to checking if a key existed before I checked it value, I forgot that if the key didn't exist then there were no neighbours!

Part 2 just added a loop to the code for part 2.  Wondered if it would be slow, but still ~100msec for 43 iterations.